Use stoptimesWithoutPatterns also for stop page
